+++ client/packages/invoices/src/InboundShipment/DetailView/modals/InboundLineEdit/TabTables.tsx
@@ -22,6 +22,8 @@ import {
   useFormatNumber,
   NumUtils,
   TextInputCell,
+  IconButton,
+  DeleteIcon,
 } from '@openmsupply-client/common';
 import { DraftInboundLine } from '../../../../types';
 import {
@@ -40,10 +42,12 @@ import {
   NumberOfPacksCell,
   vvmStatusesColumn,
 } from './utils';
+import { PatchDraftLineInput } from '../../../api';
 
 interface TableProps {
   lines: DraftInboundLine[];
-  updateDraftLine: (patch: Partial<DraftInboundLine> & { id: string }) => void;
+  updateDraftLine: (patch: PatchDraftLineInput) => void;
+  removeDraftLine?: (id: string) => void;
   isDisabled?: boolean;
   currency?: CurrencyRowFragment | null;
   isExternalSupplier?: boolean;
@@ -56,6 +60,7 @@ interface TableProps {
 export const QuantityTableComponent = ({
   lines,
   updateDraftLine,
+  removeDraftLine,
   isDisabled = false,
   hasItemVariantsEnabled,
   hasVvmStatusesEnabled,
@@ -188,6 +193,18 @@ export const QuantityTableComponent = ({
   if (displayInDoses) {
     columnDefinitions.push(...getInboundDosesColumns());
   }
+
+  columnDefinitions.push({
+    key: 'delete',
+    width: 50,
+    Cell: ({ rowData }) => (
+      <IconButton
+        label="Delete"
+        onClick={() => removeDraftLine && removeDraftLine(rowData.id)}
+        icon={<DeleteIcon fontSize="small" />}
+      />
+    ),
+  });
 
   const columns = useColumns<DraftInboundLine>(columnDefinitions, {}, [
     updateDraftLine,
